requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pinecone-client==2.2.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.0

